+++ crates/execution/src/matching_engine/engine.rs
@@ -152,7 +152,8 @@ impl OrderMatchingEngine {
         config: OrderMatchingEngineConfig,
     ) -> Self {
         let book = OrderBook::new(instrument.id(), book_type);
-        let core = OrderMatchingCore::new(instrument.id(), instrument.price_increment());
+        let mut core = OrderMatchingCore::new(instrument.id(), instrument.price_increment());
+        core.set_fill_limit_at_touch(fill_model.fill_limit_at_touch());
         let ids_generator = IdsGenerator::new(
             instrument.id().venue,
             oms_type,
@@ -370,6 +371,8 @@ impl OrderMatchingEngine {
 
     /// Sets the fill model for the matching engine.
     pub fn set_fill_model(&mut self, fill_model: FillModelAny) {
+        self.core
+            .set_fill_limit_at_touch(fill_model.fill_limit_at_touch());
         self.fill_model = fill_model;
     }